+++ core/src/main/kotlin/sift/core/jackson/NoArgConstructor.kt
@@ -1,0 +1,5 @@
+package sift.core.jackson
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class NoArgConstructor