+++ core/src/main/kotlin/sift/core/tree/EntityNode.kt
@@ -1,16 +1,32 @@
 package sift.core.tree
 
-sealed interface EntityNode {
+sealed class EntityNode(var label: String) : Comparable<EntityNode> {
+    private val properties: MutableMap<String, Any> = mutableMapOf()
+
     class Entity(
         val entity: sift.core.entity.Entity,
-        override var label: String
-    ) : EntityNode {
+        label: String
+    ) : EntityNode(label) {
         override fun toString() = label
     }
 
-    class Label(override var label: String) : EntityNode {
-        override fun toString() = label
+    class Label(label: String) : EntityNode(label)
+
+    operator fun set(property: String, value: Any) {
+        properties[property] = value
     }
 
-    var label: String
+    operator fun get(property: String): Any? = properties[property]
+
+    override fun compareTo(other: EntityNode): Int {
+        return label.compareTo(other.label)
+    }
+
+    override fun toString() = label
 }
+
+val Tree<EntityNode>.label: String
+    get() = when (val v = value) {
+        is EntityNode.Entity -> v.label
+        is EntityNode.Label  -> v.label
+    }